tests/sweep: Move sweeping, use multierror.Group
